nvim: Reconfigure gitsigns
