+++ nvim/lua/plugintmpsetup.lua
@@ -1,4 +1,3 @@
-require 'git'
 require 'orgmode-conf'
 require 'lsp-and-cmp'
 require 'dap-conf'
